requirements.txt
dnspython->2.7.0
